+++ CHANGES.rst
@@ -8,6 +8,10 @@ Changelog
 - Move tests to plone.app.testing.
   [bogdan.girman]
 
+- Supress a ZopeTestCase warning.
+  This fixes https://github.com/plone/Products.CMFPlone/issues/502
+  [gforcada]
+
 
 1.3.1 (2014-04-05)
 ------------------