+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.3.2 (unreleased)
+1.3.2 (2015-05-11)
 ------------------
 
 - Move tests to plone.app.testing.